Sprite Caching Changes
Changed the sprite retrieval so that instead of getting the sprite and storing it in memory, when the server first launches and it clearing afterwards.

It now stores all species, forms, and their sprites, in a json file in the resources, and that gets migrated into the config.

The config has also changed for discord, in-game, and settings, this is not in use yet but allows for future funcs